+++ frontend/changelog.md
@@ -1,9 +1,18 @@
 ## Changelog
 
-Tis 24-12-04
+Tis 24-12-03
 
-- Skapade github-repo och satte upp projektet. Skapade databas och en init.sql-fil. Shared-mapp och interfaces-filer för att samla alla interfaces och inte behöva ha dem i komponenterna. Gjorde endpoint för Get notes och Get categories. Skapade kommponent NoteBoard där jag började med att visa datan jag fetchat.
+- Skapade github-repo och satte upp projektet. Skapade databas och en init.sql-fil. Shared-mapp och interfaces-filer för att samla alla interfaces och inte behöva ha dem i komponenterna.
 
-Ons 24-12-05
+Ons 24-12-04
+
+- Gjorde endpoint för Get notes och Get categories. Skapade kommponent NoteBoard där jag började med att visa datan jag fetchat.
+
+Tors 24-12-05
 
 - Created e2e BDD-test that tests if its possible to open a modal and post a new note. The modal with form has not been created yet.
+
+Fre 24-12-06
+
+- Changed how to click on right element in dropdown menu in addNote.ts test.
+- Created component for button and modal with add-note form. Bdd-test is now working and i can post a new note. Created endpoint for post note.